+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # XuMiscTools Changelog
 
+## v1.2.0
+- Added LightningStrike script.
+- Improved BetterCooldownTrigger script
+- Added BetterKnife script.
+
 ## v1.1.0
 - Fixed better shovel script, improved on it a lot.